claims_system/requirements.txt
SQLAlchemy==2.0.23
PyQt6==6.6.1
python-dotenv==1.0.0
dropbox==11.36.2
pillow==10.1.0
pandas==2.1.4
cryptography==41.0.7
requests==2.31.0
pydantic==2.5.2
python-dateutil==2.8.2
pytz==2023.3.post1
bcrypt==4.1.2
